working move of the tds in guess component to new animated rows
