+++ client/src/pages/game/GuessComponent.tsx
@@ -1,7 +1,6 @@
 import React from "react";
-import { Word, WordData } from "./GamePage";
-import _isEmpty from "lodash/isEmpty";
-import { ArrowUp, ArrowDown } from "lucide-react";
+import { Word } from "./GamePage";
+import AnimatedRow from "./AnimatedRow";
 
 interface IGuessComponentProps {
   guesses: Word[];
@@ -20,74 +19,8 @@ export default function GuessComponent({
   secondaryColor,
   tertiaryColor,
 }: IGuessComponentProps) {
-
-  const compareToWordOfDay = (word: Word, index: number) => {
-    let color;
-    let numberHint;
-    if (wordOfDay) {
-      const { word_data } = word;
-      const wordOfDay_data = wordOfDay.word_data;
-      const arr1 = word_data[index].value.split(",").map((item) => item.trim());
-      const arr2 = wordOfDay_data[index].value
-        .split(",")
-        .map((item) => item.trim());
-
-      const areNumbers = (array: string[]) =>
-        array.every((item) => !isNaN(Number(item)));
-
-      const isNumericComparison = areNumbers(arr1) && areNumbers(arr2);
-
-      if (isNumericComparison) {
-        const num1 = arr1.map(Number);
-        const num2 = arr2.map(Number);
-
-        if (
-          num1.length === num2.length &&
-          num1.every((n, i) => n === num2[i])
-        ) {
-          color = primaryColor;
-        } else if (num1.some((n, i) => n > num2[i])) {
-          numberHint = "lower";
-        } else if (num1.some((n, i) => n < num2[i])) {
-          numberHint = "higher";
-        }
-      }
-
-      const matches = arr1.filter((word) => arr2.includes(word));
-
-      if (matches.length === 0) {
-        color = tertiaryColor;
-      } else if (
-        matches.length === arr1.length &&
-        matches.length === arr2.length
-      ) {
-        color = primaryColor;
-      } else {
-        color = secondaryColor;
-      }
-    }
-    return { color, numberHint };
-  };
-
-  const renderCell = (cellValue: string) => {
-    const urlRegex = /(https?:\/\/[^\s]+)/;
-    const normalizeCommas = (str: string) =>
-      str.includes(",")
-        ? str.replace(/,([^ ])/g, ", $1") // Add a space after commas without one
-        : str;
-
-    if (_isEmpty(cellValue)) {
-      return "None";
-    } else if (urlRegex.test(cellValue)) {
-      return (
-        <img src={cellValue} alt="Cell content" className="max-w-full h-auto" />
-      );
-    }
-    return normalizeCommas(cellValue);
-  };
-
   return (
-    <table className="border-separate table-fixed w-full">
+    <table className="table-fixed border-separate border-spacing-x-2 border-spacing-y-2">
       <thead>
         <tr>
           {headers.map((header: string) => (
@@ -103,34 +36,14 @@ export default function GuessComponent({
       </thead>
       <tbody>
         {guesses.map((guess: Word, index: number) => (
-          <tr key={index}>
-            {guess.word_data.map((cell: WordData, index: number) => {
-              const { color, numberHint } = compareToWordOfDay(guess, index);
-              return (
-                <td
-                  key={index}
-                  style={{
-                    backgroundColor: color,
-                  }}
-                  className="rounded-xl w-12 h-20 border text-center p-2 relative"
-                >
-                  {numberHint === "higher" && (
-                    <ArrowUp
-                      size={100}
-                      className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 text-gray-500 opacity-15"
-                    />
-                  )}
-                  {numberHint === "lower" && (
-                    <ArrowDown
-                      size={100}
-                      className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 text-gray-500 opacity-15"
-                    />
-                  )}
-                  <span className="relative">{renderCell(cell.value)}</span>
-                </td>
-              );
-            })}
-          </tr>
+          <AnimatedRow
+            key={index}
+            guess={guess}
+            wordOfDay={wordOfDay}
+            primaryColor={primaryColor}
+            secondaryColor={secondaryColor}
+            tertiaryColor={tertiaryColor}
+          />
         ))}
       </tbody>
     </table>
